[Refactor] セーブデータへの書き込みを除いてDUNGEON_IDX 型を廃止してint 型へ変えた
AVX対応をした今、セーブデータ互換性以外の箇所でビット幅をケチる価値が消滅した

+++ src/save/player-writer.cpp
@@ -185,7 +185,7 @@ void wr_player(PlayerType *player_ptr)
     wr_s16b(player_ptr->blessed);
     wr_s16b(player_ptr->tim_invis);
     wr_s16b(player_ptr->word_recall);
-    wr_s16b(player_ptr->recall_dungeon);
+    wr_s16b(static_cast<int16_t>(player_ptr->recall_dungeon));
     wr_s16b(player_ptr->alter_reality);
     wr_s16b(player_ptr->see_infra);
     wr_s16b(player_ptr->tim_infra);
